# test/e2e/__snapshots__/test_openapi/test_OpenAPIの形が変わっていないことを確認.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Speaker": {
        "description": "話者情報",
        "properties": {
          "name": {
            "title": "名前",
            "type": "string"
          },
          "speaker_uuid": {
            "title": "話者のUUID",
            "type": "string"
          },
          "styles": {
            "items": {
              "$ref": "#/components/schemas/SpeakerStyle"
            },
            "title": "スタイルの一覧",
            "type": "array"
          },
          "supported_features": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SpeakerSupportedFeatures"
              }
            ],
            "title": "話者の対応機能"
          },
          "version": {
            "default": "話者のバージョン",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "speaker_uuid",
          "styles"
        ],
        "title": "Speaker",
        "type": "object"
      },
      "SpeakerInfo": {
        "description": "話者の追加情報",
        "properties": {
          "policy": {
            "title": "policy.md",
            "type": "string"
          },
          "portrait": {
            "title": "portrait.pngをbase64エンコードしたもの",
            "type": "string"
          },
          "style_infos": {
            "items": {
              "$ref": "#/components/schemas/StyleInfo"
            },
            "title": "スタイルの追加情報",
            "type": "array"
          }
        },
        "required": [
          "policy",
          "portrait",
          "style_infos"
        ],
        "title": "SpeakerInfo",
        "type": "object"
      },
      "SpeakerStyle": {
        "description": "話者のスタイル情報",
        "properties": {
          "id": {
            "title": "スタイルID",
            "type": "integer"
          },
          "name": {
            "title": "スタイル名",
            "type": "string"
          },
          "type": {
            "default": "talk",
            "enum": [
              "talk",
              "singing_teacher",
              "frame_decode",
              "sing"
            ],
            "title": "スタイルの種類。talk:音声合成クエリの作成と音声合成が可能。singing_teacher:歌唱音声合成用のクエリの作成が可能。frame_decode:歌唱音声合成が可能。sing:歌唱音声合成用のクエリの作成と歌唱音声合成が可能。",
            "type": "string"
          }
        },
        "required": [
          "name",
          "id"
        ],
        "title": "SpeakerStyle",
        "type": "object"
      },
      "SpeakerSupportPermittedSynthesisMorphing": {
        "description": "An enumeration.",
        "enum": [
          "ALL",
          "SELF_ONLY",
          "NOTHING"
        ],
        "title": "SpeakerSupportPermittedSynthesisMorphing",
        "type": "string"
      },
      "SpeakerSupportedFeatures": {
        "description": "話者の対応機能の情報",
        "properties": {
          "permitted_synthesis_morphing": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SpeakerSupportPermittedSynthesisMorphing"
              }
            ],
            "default": "ALL",
            "title": "モーフィング機能への対応"
          }
        },
        "title": "SpeakerSupportedFeatures",
        "type": "object"
      },
      "StyleInfo": {
        "description": "スタイルの追加情報",
        "properties": {
          "icon": {
            "title": "当該スタイルのアイコンをbase64エンコードしたもの",
            "type": "string"
          },
          "id": {
            "title": "スタイルID",
            "type": "integer"
          },
          "portrait": {
            "title": "当該スタイルのportrait.pngをbase64エンコードしたもの",
            "type": "string"
          },
          "voice_samples": {
            "items": {
              "type": "string"
            },
            "title": "voice_sampleのwavファイルをbase64エンコードしたもの",
            "type": "array"
          }
        },
        "required": [
          "id",
          "icon",
          "voice_samples"
        ],
        "title": "StyleInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "DUMMY の音声合成エンジンです。",
    "title": "DUMMY Engine",
    "version": "latest"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/initialize_speaker": {
      "post": {
        "description": "指定されたスタイルを初期化します。\n実行しなくても他のAPIは使用できますが、初回実行時に時間がかかることがあります。",
        "operationId": "initialize_speaker_initialize_speaker_post",
        "parameters": [
          {
            "in": "query",
            "name": "speaker",
            "required": true,
            "schema": {
              "title": "Speaker",
              "type": "integer"
            }
          },
          {
            "description": "既に初期化済みのスタイルの再初期化をスキップするかどうか",
            "in": "query",
            "name": "skip_reinit",
            "required": false,
            "schema": {
              "default": false,
              "description": "既に初期化済みのスタイルの再初期化をスキップするかどうか",
              "title": "Skip Reinit",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Initialize Speaker",
        "tags": [
          "その他"
        ]
      }
    },
...
    "/is_initialized_speaker": {
      "get": {
        "description": "指定されたスタイルが初期化されているかどうかを返します。",
        "operationId": "is_initialized_speaker_is_initialized_speaker_get",
        "parameters": [
          {
            "in": "query",
            "name": "speaker",
            "required": true,
            "schema": {
              "title": "Speaker",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Is Initialized Speaker Is Initialized Speaker Get",
                  "type": "boolean"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Is Initialized Speaker",
        "tags": [
          "その他"
        ]
      }
    },
...
    "/singer_info": {
      "get": {
        "description": "指定されたspeaker_uuidの歌手に関する情報をjson形式で返します。\n画像や音声はbase64エンコードされたものが返されます。",
        "operationId": "singer_info_singer_info_get",
        "parameters": [
          {
            "in": "query",
            "name": "speaker_uuid",
            "required": true,
            "schema": {
              "title": "Speaker Uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeakerInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Singer Info",
        "tags": [
          "その他"
        ]
      }
    },
    "/singers": {
      "get": {
        "description": "歌手情報の一覧を取得します",
        "operationId": "singers_singers_get",
        "parameters": [
          {
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Speaker"
                  },
                  "title": "Response Singers Singers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Singers",
        "tags": [
          "その他"
        ]
      }
    },
    "/speaker_info": {
      "get": {
        "description": "指定されたspeaker_uuidの話者に関する情報をjson形式で返します。\n画像や音声はbase64エンコードされたものが返されます。",
        "operationId": "speaker_info_speaker_info_get",
        "parameters": [
          {
            "in": "query",
            "name": "speaker_uuid",
            "required": true,
            "schema": {
              "title": "Speaker Uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeakerInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Speaker Info",
        "tags": [
          "その他"
        ]
      }
    },
    "/speakers": {
      "get": {
        "description": "話者情報の一覧を取得します。",
        "operationId": "speakers_speakers_get",
        "parameters": [
          {
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Speaker"
                  },
                  "title": "Response Speakers Speakers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Speakers",
        "tags": [
          "その他"
        ]
      }
    },
...
  }
}